fix: suppress websocket error after connection lost (#3270)
* report bug template add `Electron`

* mute LSP WebSocket connection lost pop up error

* refine re-connection

* use SendRequestWithRetry
