+++ Composer/packages/lib/code-editor/src/LgEditor.tsx
@@ -1,26 +1,21 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { listen, MessageConnection } from 'vscode-ws-jsonrpc';
 import get from 'lodash/get';
-import { MonacoServices, MonacoLanguageClient } from 'monaco-languageclient';
+import { MonacoServices } from 'monaco-languageclient';
 import { EditorDidMount } from '@monaco-editor/react';
 
 import { registerLGLanguage } from './languages';
-import { createUrl, createWebSocket, createLanguageClient } from './utils/lspUtil';
+import { createUrl, createWebSocket, createLanguageClient, SendRequestWithRetry } from './utils/lspUtil';
 import { BaseEditor, BaseEditorProps, OnInit } from './BaseEditor';
+import { LGOption } from './utils';
 
 const LG_HELP =
   'https://github.com/microsoft/BotBuilder-Samples/blob/master/experimental/language-generation/docs/lg-file-format.md';
 const placeholder = `> To learn more about the LG file format, read the documentation at
 > ${LG_HELP}`;
-
-export interface LGOption {
-  projectId?: string;
-  fileId: string;
-  templateId?: string;
-}
 
 export interface LGLSPEditorProps extends BaseEditorProps {
   lgOption?: LGOption;
@@ -40,15 +35,6 @@ const defaultLGServer = {
 declare global {
   interface Window {
     monacoServiceInstance: MonacoServices;
-    monacoLGEditorInstance: MonacoLanguageClient;
-  }
-}
-
-async function initializeDocuments(lgOption: LGOption | undefined, uri: string) {
-  const languageClient = window.monacoLGEditorInstance;
-  if (languageClient) {
-    await languageClient.onReady();
-    languageClient.sendRequest('initializeDocuments', { uri, lgOption });
   }
 }
 
@@ -68,6 +54,29 @@ export function LgEditor(props: LGLSPEditorProps) {
     editorId = [projectId, fileId, templateId].join('/');
   }
 
+  const [editor, setEditor] = useState<any>();
+
+  useEffect(() => {
+    if (!editor) return;
+
+    if (!window.monacoServiceInstance) {
+      window.monacoServiceInstance = MonacoServices.install(editor as any);
+    }
+
+    const uri = get(editor.getModel(), 'uri._formatted', '');
+    const url = createUrl(lgServer);
+    const webSocket: WebSocket = createWebSocket(url);
+    listen({
+      webSocket,
+      onConnection: (connection: MessageConnection) => {
+        const languageClient = createLanguageClient('LG Language Client', ['botbuilderlg'], connection);
+        SendRequestWithRetry(languageClient, 'initializeDocuments', { lgOption, uri });
+        const disposable = languageClient.start();
+        connection.onClose(() => disposable.dispose());
+      },
+    });
+  }, [editor]);
+
   const onInit: OnInit = (monaco) => {
     registerLGLanguage(monaco);
 
@@ -77,31 +86,7 @@ export function LgEditor(props: LGLSPEditorProps) {
   };
 
   const editorDidMount: EditorDidMount = (_getValue, editor) => {
-    if (!window.monacoServiceInstance) {
-      window.monacoServiceInstance = MonacoServices.install(editor as any);
-    }
-
-    if (!window.monacoLGEditorInstance) {
-      const uri = get(editor.getModel(), 'uri._formatted', '');
-      const url = createUrl(lgServer);
-      const webSocket: WebSocket = createWebSocket(url);
-      listen({
-        webSocket,
-        onConnection: (connection: MessageConnection) => {
-          const languageClient = createLanguageClient('LG Language Client', ['botbuilderlg'], connection);
-          if (!window.monacoLGEditorInstance) {
-            window.monacoLGEditorInstance = languageClient;
-          }
-          initializeDocuments(lgOption, uri);
-          const disposable = languageClient.start();
-          connection.onClose(() => disposable.dispose());
-        },
-      });
-    } else {
-      const uri = get(editor.getModel(), 'uri._formatted', '');
-      initializeDocuments(lgOption, uri);
-    }
-
+    setEditor(editor);
     if (typeof props.editorDidMount === 'function') {
       return props.editorDidMount(_getValue, editor);
     }
