+++ Composer/packages/lib/code-editor/src/utils/index.ts
@@ -2,3 +2,4 @@
 // Licensed under the MIT License.
 
 export * from './common';
+export * from './types';
