+++ Composer/packages/lib/code-editor/src/utils/types.ts
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export interface LGOption {
+  projectId?: string;
+  fileId: string;
+  templateId?: string;
+}
+
+export interface LUOption {
+  projectId?: string;
+  fileId: string;
+  sectionId?: string;
+}
